@@ -1,10 +1,13 @@
 """
-File utilities
+Convenience functions related to files and file system
 """
 
 import os
 import platform
+import shlex
+
 from pathlib import Path, PurePath
+
 
 def findFileInPath(element, dirPath=None):
     """
@@ -26,7 +29,7 @@ def findFileInPath(element, dirPath=None):
     filesFound = []
 
     if dirPath is None:
-        dirPath = os.getenv('PATH')
+        dirPath = os.getenv("PATH")
 
     if isinstance(dirPath, str):
         dirs = dirPath.split(os.pathsep)
@@ -41,11 +44,9 @@ def findFileInPath(element, dirPath=None):
     return filesFound
 
 
-def getFileList(strPath,
-                wildcard=None,
-                fullpath=False,
-                recursive=False,
-                strName=False):
+def getFileList(
+    strPath, wildcard="*.*", fullpath=False, recursive=False, strName=False
+):
     """
     Get files in a directory
     strPath has to be an existing directory or file
@@ -53,7 +54,33 @@ def getFileList(strPath,
     strExtFilter in the form -> .ext
     """
 
+    doubleWildcard = False
+    argWildcard = wildcard
+    lstObjFileNames = []
+
     p = Path(strPath)
+
+    try:
+        if p.is_file():
+            p = p.parent
+    except OSError:
+        pass
+
+    try:
+        if not p.is_dir():
+            p = p.parent
+    except OSError:
+        wildcard = p.stem
+        p = p.parent
+        doubleWildcard = True  # wildcard found on strPath argument
+
+    try:
+        if not p.is_dir():
+            # Wrong argument for strPath
+            return []
+    except OSError:
+        # Wrong argument for strPath
+        return []
 
     if (not p.is_file()) and (not p.is_dir()):
         return []
@@ -63,15 +90,103 @@ def getFileList(strPath,
     if p.is_file():
         p = p.parent
 
-    if wildcard is None:
-        wc = "*.*"
+    wc = wildcard
+    awc = argWildcard
+
+    if doubleWildcard:
+
+        lstFiles = [x for x in p.glob(wc) if x.is_file()]
+
+        if recursive:
+            awc = "**/" + stripEncaseQuotes(argWildcard)
+
+            for d in lstFiles:
+                if d.is_dir():
+                    oFiles = [x for x in d.glob(awc) if x.is_file()]
+                    lstObjFileNames.extend(oFiles)
+        else:
+            lstObjFileNames.extend(lstFiles)
+
     else:
-        wc = wildcard
+        if recursive:
+            wc = "**/" + stripEncaseQuotes(wildcard)
 
-    if recursive:
-        wc = "**/" + stripEncaseQuotes(wildcard)
+        # print("Path = {}\nWith wildcard ={}\n".format(p, wc))
 
-    lstObjFileNames = [x for x in p.glob(wc) if x.is_file()]
+        lstObjFileNames = [x for x in p.glob(wc) if x.is_file()]
+
+    if not fullpath:
+        lstFilesFilter = [x.name for x in lstObjFileNames]
+        return lstFilesFilter
+
+    if strName:
+        lstFilesFilter = [str(x) for x in lstObjFileNames]
+        return lstFilesFilter
+
+    return lstObjFileNames
+
+
+def getDirectoryList(
+    strPath, wildcard="*", fullpath=False, recursive=False, strName=False
+):
+    """
+    Get files in a directory
+    strPath has to be an existing directory or file
+    in case of a file the parent directory is used
+    strExtFilter in the form -> .ext
+    """
+
+    doubleWildcard = False
+    argWildcard = wildcard
+    lstObjFileNames = []
+
+    p = Path(strPath)
+
+    try:
+        if p.is_file():
+            p = p.parent
+    except OSError:
+        pass
+
+    try:
+        if not p.is_dir():
+            p = p.parent
+    except OSError:
+        wildcard = p.stem
+        p = p.parent
+        doubleWildcard = True  # wildcard found on strPath argument
+
+    try:
+        if not p.is_dir():
+            # Wrong argument for strPath
+            return []
+    except OSError:
+        # Wrong argument for strPath
+        return []
+
+    lstFilesFilter = []
+
+    wc = wildcard
+    awc = argWildcard
+
+    if doubleWildcard:
+        dirs = [x for x in p.glob(wc) if x.is_dir()]
+
+        if recursive:
+            awc = "**/" + stripEncaseQuotes(argWildcard)
+            for d in dirs:
+                if d.is_dir():
+                    lstObjFileNames.append(d)
+                    oDirs = [x for x in d.glob(awc) if x.is_dir()]
+                    lstObjFileNames.extend(oDirs)
+
+        else:
+            lstObjFileNames.extend(dirs)
+    else:
+        if recursive:
+            wc = "**/" + stripEncaseQuotes(wildcard)
+
+        lstObjFileNames = [x for x in p.glob(wc) if x.is_dir()]
 
     if not fullpath:
         lstFilesFilter = [x.name for x in lstObjFileNames]
@@ -95,43 +210,31 @@ def getExecutable(search):
 
     if currentOS == "Darwin":
 
-        lstTest = Path("/Applications").glob('**/' + fileToSearch)
+        lstTest = Path("/Applications").glob("**/" + fileToSearch)
 
         for l in lstTest:
-            p = Path(l)
-            if p.stem == fileToSearch:
+            if (p := Path(l)).stem == fileToSearch:
                 return p
 
     elif currentOS == "Windows":
 
-        if fileToSearch.find('.') < 0:
+        if fileToSearch.find(".") < 0:
             # assume is binary executable
-            fileToSearch += '.exe'
-
-        defPrograms64 = os.environ.get('ProgramFiles')
-        defPrograms32 = os.environ.get('ProgramFiles(x86)')
+            fileToSearch += ".exe"
 
         dirs = []
-        if defPrograms64 is not None:
-            dirs.append(defPrograms64)
-
-        if defPrograms32 is not None:
-            dirs.append(defPrograms32)
+        dirs.append(os.environ.get("ProgramFiles") or "")
+        dirs.append(os.environ.get("ProgramFiles(x86)") or "")
 
         # search 64 bits
         for d in dirs:
-            search = sorted(Path(d).rglob(fileToSearch))
-            if search:
-                executable = Path(search[0])
-                if executable.is_file():
+            if search := sorted(Path(d).rglob(fileToSearch)):
+                if (executable := Path(search[0])).is_file():
                     return executable
 
-    searchFile = findFileInPath(fileToSearch)
-
-    if searchFile:
+    if searchFile := findFileInPath(fileToSearch):
         for e in searchFile:
-            executable = Path(e)
-            if executable.is_file():
+            if (executable := Path(e)).is_file():
                 return executable
 
     return None
@@ -156,5 +259,13 @@ def stripEncaseQuotes(strFile):
 
     if (s[0:1] == "'") and (s[-1:] == "'"):
         s = s[1:-1]
+
+    return s
+
+
+def fileQuote(oFile):
+
+    f = shlex.quote(str(oFile))
+    s = f.replace('"\'"', r"\'")
 
     return s
