@@ -2,5 +2,22 @@
 
 # MKV0001
 
-from .classes import MKVCommand, VerifyStructure, VerifyMKVCommand
-from .mkvutils import getMKVMerge, getMKVMergeVersion, stripEncaseQuotes
+from .classes import (
+    MKVAttachment,
+    MKVAttachments,
+    MKVCommand,
+    MKVCommandParser,
+    MKVCommandNew,
+    SourceFile,
+    SourceFiles,
+    VerifyMKVCommand,
+    VerifyStructure,
+)
+from .mkvutils import (
+    getMKVMerge,
+    getMKVMergeVersion,
+    numberOfTracksInCommand,
+    resolveOverwrite,
+    stripEncaseQuotes,
+    unQuote,
+)
