@@ -5,19 +5,17 @@ this information is used for internationalization
 
 from PySide2.QtWidgets import QAction
 
+
 class QActionWidget(QAction):
     """
-    class QAction([parent=None])¶
-    QAction(icon, text[, parent=None])
-    QAction(text[, parent=None])
-    """
-    def __init__(self,
-                 *args,
-                 shortcut=None,
-                 tooltip=None,
-                 **kwargs
-                ):
+    QActionWidget subclass of QAction save original shortcut and tooltip for locale application
 
+    Args:
+        shortcut (str, optional): original shortcut string representation. Defaults to None.
+        tooltip (str, optional): original tooltip. Defaults to None.
+    """
+
+    def __init__(self, *args, shortcut=None, tooltip=None, **kwargs):
         super().__init__(*args, **kwargs)
 
         for p in args:
@@ -29,17 +27,20 @@ class QActionWidget(QAction):
 
         if shortcut is not None:
             self.setShortcut(shortcut)
+
         if tooltip is not None:
-            self.setStatusTip(tooltip)
+            self.setToolTip(tooltip)
 
     def setShortcut(self, shortcut, *args, **kwargs):
 
         if self.shortcut is None:
             self.shortcut = shortcut
-        super().setStatusTip(shortcut, *args, **kwargs)
 
-    def setStatusTip(self, tooltip, *args, **kwargs):
+        super().setShortcut(shortcut, *args, **kwargs)
+
+    def setToolTip(self, tooltip, *args, **kwargs):
 
         if self.tooltip is None:
             self.tooltip = tooltip
-        super().setStatusTip(tooltip, *args, **kwargs)
+
+        super().setToolTip(tooltip, *args, **kwargs)
